+++ posthog-react-native/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+# 2.10.2 - 2024-01-22
 
 1. Do not try to load the `expo-file-system` package on the Web target since it's not supported.
 2. if `react-native-device-info` is available for the Web target, do not set `unknown` for all properties.